Move tests to iceberg-spark
