+++ settings.gradle
@@ -22,7 +22,6 @@ include 'api'
 include 'common'
 include 'core'
 include 'data'
-include 'examples'
 include 'orc'
 include 'parquet'
 include 'spark'
