+++ spark/src/test/java/org/apache/iceberg/examples/data/books.json
