+++ spark/src/test/java/org/apache/iceberg/examples/data/new-books.json
